1981.2: Move `barItems`
