+++ administration/src/bp-modules/applications/constants/index.ts
@@ -12,29 +12,6 @@ export type ApplicationStatusBarItemType = {
   status?: ApplicationStatus
 }
 
-export const barItems: ApplicationStatusBarItemType[] = [
-  {
-    title: 'allApplications',
-    status: undefined,
-  },
-  {
-    title: 'accepted',
-    status: ApplicationStatus.fullyVerified,
-  },
-  {
-    title: 'rejected',
-    status: ApplicationStatus.fullyRejected,
-  },
-  {
-    title: 'withdrawed',
-    status: ApplicationStatus.withdrawed,
-  },
-  {
-    title: 'open',
-    status: ApplicationStatus.ambiguous,
-  },
-]
-
 export enum VerificationStatus {
   Verified,
   Rejected,
